# crc: optimize by supporting arm xor fusion feature