@pablodelara 2023-08-11
Hi @liuqinfei. Could you modify the signed-off and use your full name, as per our contribution guidelines?
https://01.org/community/signed-process. Thanks!