@liuqinfei 2023-08-14
> Hi @liuqinfei. Could you modify the signed-off and use your full name, as per our contribution guidelines? https://01.org/community/signed-process. Thanks!

Oh, in fact. liuqinfei is my full name. liu is my last name. qinfei is my first name. Thanks.